test: 픽스쳐 클래스를 사용하여 테스트를 용이하게 하도록 수정
