+++ src/main/java/janggi/dto/PositionInfo.java
@@ -13,14 +13,6 @@ public record PositionInfo(
         Piece piece,
         Point point
 ) {
-    public static PositionInfo from(List<String> data) {
-        return from(
-                Team.valueOf(data.get(0)),
-                data.get(1),
-                Integer.parseInt(data.get(2)),
-                Integer.parseInt(data.get(3))
-        );
-    }
 
     public static PositionInfo from(Team team, String pieceName, int x, int y) {
         return new PositionInfo(
