+++ src/test/java/janggi/domain/piece/PhoTest.java
@@ -7,6 +7,7 @@ import janggi.domain.Board;
 import janggi.domain.Point;
 import janggi.domain.status.Team;
 import janggi.dto.PositionInfo;
+import janggi.fixture.PositionInfoFixture;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -23,13 +24,13 @@ public class PhoTest {
     void setUp() {
         board = new Board();
         List<PositionInfo> info = new ArrayList<>();
-        info.add(PositionInfo.from(List.of("HAN", "PHO", "1", "1")));
-        info.add(PositionInfo.from(List.of("HAN", "CHA", "1", "2")));
-        info.add(PositionInfo.from(List.of("HAN", "CHA", "1", "3")));
-        info.add(PositionInfo.from(List.of("CHO", "PHO", "1", "5")));
-        info.add(PositionInfo.from(List.of("CHO", "PHO", "3", "0")));
-        info.add(PositionInfo.from(List.of("CHO", "JANG", "4", "1")));
-        info.add(PositionInfo.from(List.of("CHO", "PHO", "1", "6")));
+        info.add(PositionInfoFixture.from(List.of("HAN", "PHO", "1", "1")));
+        info.add(PositionInfoFixture.from(List.of("HAN", "CHA", "1", "2")));
+        info.add(PositionInfoFixture.from(List.of("HAN", "CHA", "1", "3")));
+        info.add(PositionInfoFixture.from(List.of("CHO", "PHO", "1", "5")));
+        info.add(PositionInfoFixture.from(List.of("CHO", "PHO", "3", "0")));
+        info.add(PositionInfoFixture.from(List.of("CHO", "JANG", "4", "1")));
+        info.add(PositionInfoFixture.from(List.of("CHO", "PHO", "1", "6")));
         board.init(info);
     }
