+++ src/test/java/janggi/domain/status/HanTurnTest.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 
 import janggi.domain.Board;
 import janggi.domain.Point;
-import janggi.domain.piece.Piece;
 import janggi.dto.PositionInfo;
+import janggi.fixture.PositionInfoFixture;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -21,10 +21,10 @@ public class HanTurnTest {
     void setUp() {
         board = new Board();
         List<PositionInfo> info = new ArrayList<>();
-        info.add(PositionInfo.from(List.of("HAN","JANG", "4", "1")));
-        info.add(PositionInfo.from(List.of("CHO","JANG", "4", "8")));
-        info.add(PositionInfo.from(List.of("HAN", "CHA", "1", "1")));
-        info.add(PositionInfo.from(List.of("CHO", "CHA", "2", "3")));
+        info.add(PositionInfoFixture.from(List.of("HAN","JANG", "4", "1")));
+        info.add(PositionInfoFixture.from(List.of("CHO","JANG", "4", "8")));
+        info.add(PositionInfoFixture.from(List.of("HAN", "CHA", "1", "1")));
+        info.add(PositionInfoFixture.from(List.of("CHO", "CHA", "2", "3")));
         board.init(info);
     }
 
@@ -49,7 +49,7 @@ public class HanTurnTest {
         // given
         Board board = new Board();
         List<PositionInfo> info = new ArrayList<>();
-        info.add(PositionInfo.from(List.of("CHO", "CHA", "1", "1")));
+        info.add(PositionInfoFixture.from(List.of("CHO", "CHA", "1", "1")));
         board.init(info);
         Point from = Point.of(1, 1);
         Point to = Point.of(2, 3);
